add interface online/offline

+++ feiapi-backend/src/main/java/com/fei/project/common/PageRequest.java
@@ -6,8 +6,7 @@ import lombok.Data;
 /**
  * 分页请求
  *
- * @author <a href="https://github.com/">fei</a>
- * 
+ * @author fei
  */
 @Data
 public class PageRequest {
@@ -15,12 +14,12 @@ public class PageRequest {
     /**
      * 当前页号
      */
-    private int current = 1;
+    private long current = 1;
 
     /**
      * 页面大小
      */
-    private int pageSize = 10;
+    private long pageSize = 10;
 
     /**
      * 排序字段
